+++ package.json
@@ -31,6 +31,7 @@
     "react-native-otp-textinput": "^1.1.6",
     "react-native-paper": "^5.12.3",
     "react-native-paper-dropdown": "^1.0.7",
+    "react-native-root-siblings": "^5.0.1",
     "react-native-safe-area-context": "^4.10.1",
     "react-native-screens": "^3.31.1",
     "react-native-sound-player": "^0.14.2",
